# Conflicts:
#	Springboot/src/main/java/com/centralstudenthub/entity/TeachingStaffContact.java
#	Springboot/src/main/java/com/centralstudenthub/entity/student_profile/Warning.java
#	Springboot/src/main/java/com/centralstudenthub/service/AssignmentService.java

+++ Springboot/src/main/java/com/centralstudenthub/Model/Request/StudentAssignmentAnswerRequest.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class StudentAssignmentAnswerRequest {
     private String answerPath;
-    private Double grade;
-    private Long studentProfileId;
+    private double grade;
+    private int studentProfileId;
     private Long assignmentId;
 }
